Adjust indentation of HTML files
I’ve been more generous with indentation in later files – let’s do the
same for earlier files and the index as well, for consistency.

Annoyingly, this is not quite the indentation which Emacs produces
unassisted, because Emacs doesn’t know that <source> is an empty element
without an end tag (at least not in the version installed on my server).