Add RSS file
The RSS file is generated from the RDF description using a script I
wrote for this purpose, which is cloned into my personal git directory.
(Perhaps I’ll install it into ~/.local/bin or /usr/local/bin at some
point so the absolute path isn’t necessary, dunno.)

In the Makefile, the RSS file depends on the current directory itself,
which means that `make` will regenerate it if any entries (such as
subdirectories for new posts) were added to the directory. (It will also
regenerate the file if anything was removed.)

I’m also adding the RSS file itself, even though it’s generated content,
since the generation is not (currently) deterministic (the lastBuildDate
is always the current date), it’s fairly small, and it can take a while
(several seconds) to generate, especially if you’re not running the
script on the same server that also serves the blog.