Add time to two recordings on the same day
So they don’t get reordered in index.rss when I rebuild it. The times
are taken from the tweets about the posts [1][2] and should be
reasonably close to the actual publication time, I think. (They’re
specified without a time zone offset because otherwise rdf2rss crashes
when trying to compare them to the other times.)

[1]: https://twitter.com/LucasWerkmeistr/status/891358124044668929
[2]: https://twitter.com/LucasWerkmeistr/status/891395234395021316